+++ ApolloState.m
@@ -17,5 +17,7 @@ NSInteger sReadPostMaxCount = 0;
 
 NSInteger sUnmuteCommentsVideos = 0; // 0=Default, 1=Remember from Full Screen, 2=Always
 
+BOOL sProxyImgurDDG = NO;
+
 NSString *sPushNotificationServer = nil;
 NSString *sServerToken = nil;